TECH-5691 : 블록체인 새 상품 등록

+++ src/main/java/io/temco/guhada/blockchain/service/TrackRecordService.java
@@ -6,9 +6,19 @@ import io.temco.guhada.blockchain.model.request.TrackRecord;
 
 public interface TrackRecordService {
 	
+	/**
+	 * upload new deal data to blockchain.  
+	 * 
+	 * @param trackRecordRequest
+	 */
 	void uploadProductInfo(TrackRecord trackRecordRequest);
 	
-	List<TrackRecord> getProductTransactions(int dealId);
+	/**
+	 * get deal data from  blockchain.  
+	 * 
+	 * @param dealId : deal id
+	 */
+	List<TrackRecord> getProductTransactions(Long dealId);
 	
 	void testUpload();
 }
